+++ src/Contract/ConfigInterface.php
@@ -14,6 +14,14 @@ interface ConfigInterface
     public function addNamespace(string $name, $config): ConfigInterface;
 
     /**
+     * Add an override for the config such as cli argument or environment variables
+     *
+     * @param OverrideInterface $override
+     * @return ConfigInterface
+     */
+    public function addOverride(OverrideInterface $override): ConfigInterface;
+
+    /**
      * Get the config value or return the default value if none is set
      *
      * @param string $name