+++ src/Contract/OverrideInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Mizmoz\Config\Contract;
+
+/**
+ * Override the config values
+ *
+ * @package Mizmoz\Config\Contract
+ */
+interface OverrideInterface
+{
+    /**
+     * Override the values in the config
+     *
+     * @param ConfigInterface $config
+     * @return ConfigInterface
+     */
+    public function override(ConfigInterface $config): ConfigInterface;
+}